Do not scan a file which cannot be read
When the media scanner is invoked to scan a file which it cannot access
(e.g. one in the app-private directories) the related database entry
will be created anyway, which doesn't make sense. This results in having
broken entries showed up in e.g. Album app.

This fix is to prevent scanning files which are inaccessible.

Change-Id: I5b4909bf709c82d66e891f3e7f6890febc90c6eb
Signed-off-by: reS28raM <paolo.a.ulivi@gmail.com>