NullPointerException when starting VoiceInteractionManagerService
If an end user disables the package that implements support for
Voice Assist (Google App) and reboots, an NPE is thrown during boot,
which results in the device getting stuck at the boot animation.

Recovery is not possible without USB debugging enabled, leaving factory
reset/user data wipe as the only other option.

Change-Id: I9c161fff7f096deee13bf8e37d951e20898ae3ff
Signed-off-by: reS28raM <paolo.a.ulivi@gmail.com>